Add explicit [Any] to ColumnElement in Generic.
SQLAlchemy now defines ColumnElement as generic, and we don't really
want to make use of that since we usually hold columns in heterogeneous
containers (I imagine this typing is much more useful for the ORM).

In most contexts, not providing a type variable is interpreted as
ColumnElement[Any], but it seems to be causing trouble here in
inheritance involving Generic, as it makes Python want the Generic to
include the typevar for ColumnElement.